Add stress.jar to rpmbuild
Summary: We are adding the stress tool library to the rpmbuild so that it can be easily deployed through chef/yum. We will deploy specifically this version of Cassandra, along with the scripts that use this library, to the benchmark machines.

Test Plan:
Built the rpmbuild and checked the file contents to confirm that the `stress.jar` is a part of the build as so:
```
$ rpm -qlp apache-cassandra-2.2.5+git20160711.15b57fd-1.noarch.rpm
...
/usr/share/cassandra/lib/stress.jar
...
```
Also locally installed this build on my dev machine as so:
```
$ yum reinstall apache-cassandra-2.2.5+git20160711.15b57fd-1.noarch.rpm
```
Then checked that `/usr/share/cassandra/lib` does indeed contain the `stress.jar`:
```
$ ls /usr/share/cassandra/lib
...
stress.jar
...
```

Reviewers: #ig-cassandra, awhang, dikang, chenshen

Reviewed By: chenshen

Differential Revision: https://phabricator.intern.facebook.com/D3700212

Tasks: 12626959

Signature: t1:3700212:1470868823:3170f831e62e023c51cbef192283a4455bac64e0